wip - spike on refactoring internal packages to be smaller
