--- senders/configs.go
@@ -1,1 +1,0 @@
-package senders
